Started writing out perturbations to universe
On reflection, it would be nicer to work in a gauge invariant manner
from the beginning a-la-Mukhanov. However, he works in conformal time,
which I think makes things less clear. It would be nice to present a
combination of Mukhanov and Baummann. This may be too much extra work,
but then I could quite quickly derive some physics for the kinetically
dominated universe...